Created ability to adjust window size. Game elements are scaled to fit the screen.
Work is still needed to fix the title screen, which scales improperly. Also the top of the screen can be cut off, which shouldn't happen. Also, Need to draw border over range to clarify what part of the bottom is extra, and to cover the sides so the player can't see more of the screen than they're supposed to.