Fixed enemy movement on top of moving walls.
This was done by giving enemies a relative velocity, as well as a pointer to a moving wall, and telling them to update their velocity and every game update, similar to the player. The problem with this is, that this prevents the feature of being able to push an enemy and have them infinitely fly until they collide with a wall. May or may not re-add this in the future.